header js refactored into own module
