+++ assets/js/accounts.js
@@ -1,3 +1,4 @@
 import "../css/accounts.scss";
+import "./common/header";
 import { setup } from "./live_view_boilerplate";
 setup({});
