+++ assets/js/app.js
@@ -1,22 +1,4 @@
 import "../css/app.scss";
+import "./common/header";
 import { setup } from "./live_view_boilerplate";
 setup({});
-
-window.toggle_user_profile_dropdown = function () {
-  const element = document.querySelector(".header .user-profile-dropdown");
-  element.classList.toggle("active");
-};
-
-window.addEventListener("click", (event) => {
-  if (
-    event.target.closest(".user-profile-dropdown") ||
-    event.target.closest(".user-profile")
-  ) {
-    return;
-  }
-
-  const element = document.querySelector(".header .user-profile-dropdown");
-  if (element) {
-    element.classList.remove("active");
-  }
-});
