+++ assets/js/contestant.js
@@ -1,4 +1,5 @@
 import "../css/contestant.scss";
+import "./common/header";
 
 import { setup } from "./live_view_boilerplate";
 import DrawName from "./hooks/draw_name";
