+++ assets/js/trebek.js
@@ -1,3 +1,4 @@
 import "../css/trebek.scss";
+import "./common/header";
 import { setup } from "./live_view_boilerplate";
 setup({});
