+++ assets/js/tv.js
@@ -1,4 +1,5 @@
 import "../css/tv.scss";
+import "./common/header";
 
 import { setup } from "./live_view_boilerplate";
 import Stats from "./hooks/stats";
